libs/AI Agent 与大模型应用开发实战手册/第七章 垂直场景实战项目/7.2 项目二：企业知识库智能问答/requirement.txt
markitdown[all]==0.1.1
qdrant-client==1.9.1
fastembed==0.3.3
rank-bm25==0.2.2
sentence-transformers==3.0.1
openai==1.35.0
langchain-text-splitters==0.2.2
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.7.4
fastapi==0.111.0
uvicorn==0.30.1
tqdm==4.66.4
diskcache==5.6.3